Lots of changes check desc.
Missiles blow up when hitting each other, missiles have forces when they go off, added enemy boat that takes damage to rocks/missiles/bombs.  Added projectiles that either go through the water like cannon balls, or float on the water and then detonate after 3 seconds.  Added cloud enemy that either goes back and forth on screen endlessly until hit by player or missile or can be set to a mode where it tries to always hover directly over the player.  Started working on Kraken, and even though the basics are there, there is still much work to be done.  Finally, added a small thing where if the boat is told that it reached the end of the level it will automatically sail off screen and then it will fade to black.